fix(clipboard): route all writeText callers through the non-secure-context helper
`navigator.clipboard` is undefined outside a secure context (https,
localhost, 127.0.0.1). When Kolu is reached over plain http to a LAN
address, hostname, or Tailscale IP — typical on Windows/Chrome — direct
`navigator.clipboard.writeText` calls throw `TypeError: Cannot read
properties of undefined (reading 'writeText')`, with no permission prompt
to recover from. The helper in `terminal/clipboard.ts` already handled
this via a hidden-textarea + `document.execCommand("copy")` fallback, but
four call sites bypassed it: CommentsTray (the "Copy to clipboard" for
queued comments — the user-reported repro), DiagnosticInfo, the code
context menu, and the Pierre file-tree "Copy path".

Move the helper to `ui/clipboard.ts` (it's no longer terminal-specific
now that the whole UI uses it) and expand the module-level docs to
cover: why execCommand is the only portable escape hatch
(caniuse 100/100, no removal timeline), the user-gesture requirement,
the rejection fallthrough, the OSC-52 read fallback gap, and HTTPS as
the long-term cure.
