+++ packages/client/src/comments/CommentsTray.tsx
@@ -8,6 +8,7 @@
 
 import { type Component, createMemo, For, Show } from "solid-js";
 import { toast } from "solid-sonner";
+import { writeTextToClipboard } from "../ui/clipboard";
 import { formatMarkdown } from "./formatMarkdown";
 import type { Comment } from "./types";
 import { useComments } from "./useComments";
@@ -31,7 +32,7 @@ export const CommentsTray: Component<CommentsTrayProps> = (props) => {
     if (list.length === 0) return;
     const text = formatMarkdown(list);
     try {
-      await navigator.clipboard.writeText(text);
+      await writeTextToClipboard(text);
       toast.success(
         `Copied ${list.length} comment${list.length === 1 ? "" : "s"} to clipboard`,
       );
