+++ packages/client/src/ui/CodeContextMenu.tsx
@@ -8,8 +8,8 @@
 
 import { type Component, createSignal, For, onCleanup, Show } from "solid-js";
 import { Dynamic, Portal } from "solid-js/web";
-import { toast } from "solid-sonner";
 import { match } from "ts-pattern";
+import { copyTextWithToast } from "./clipboard";
 
 /** Two verbs over the same selection noun: copy a string to the clipboard,
  *  or invoke an action callback. The discriminator keeps the dispatch
@@ -87,10 +87,10 @@ export const CodeContextMenu: Component<{
   const handleItem = (item: CodeContextMenuItem) => {
     match(item)
       .with({ kind: "copy" }, ({ textToCopy }) => {
-        navigator.clipboard
-          .writeText(textToCopy)
-          .then(() => toast.success(`Copied: ${textToCopy}`))
-          .catch((err: Error) => toast.error(`Failed to copy: ${err.message}`));
+        void copyTextWithToast(textToCopy, {
+          success: `Copied: ${textToCopy}`,
+          failure: "Failed to copy",
+        });
       })
       .with({ kind: "action" }, ({ onActivate }) => onActivate())
       .exhaustive();
